requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
idna==2.8
newsapi-python==0.2.3
nltk==3.4
oauthlib==3.0.1
PySocks==1.6.8
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.7.3
textblob==0.15.2
tweepy==3.7.0
urllib3==1.24.1
